requirements.txt
PyYAML
beautifulsoup4